+++ src/test/java/com/programmers/voucher/ConsoleMenuTest.java
@@ -1,9 +1,8 @@
 package com.programmers.voucher;
 
-import com.programmers.voucher.domain.customer.controller.CustomerController;
 import com.programmers.voucher.global.io.Console;
 import com.programmers.voucher.global.io.command.ConsoleCommandType;
-import com.programmers.voucher.global.io.command.CustomerCommandType;
+import com.programmers.voucher.global.io.menu.ConsoleCustomerMenu;
 import com.programmers.voucher.global.io.menu.ConsoleMenu;
 import com.programmers.voucher.global.io.menu.ConsoleVoucherMenu;
 import org.junit.jupiter.api.DisplayName;
@@ -13,7 +12,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static org.mockito.BDDMockito.*;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.then;
 
 @ExtendWith(MockitoExtension.class)
 class ConsoleMenuTest {
@@ -25,7 +25,7 @@ class ConsoleMenuTest {
     private Console console;
 
     @Mock
-    private CustomerController customerController;
+    private ConsoleCustomerMenu consoleCustomerMenu;
 
     @Mock
     private ConsoleVoucherMenu consoleVoucherMenu;
@@ -44,105 +44,16 @@ class ConsoleMenuTest {
     }
 
     @Test
-    @DisplayName("성공: customer 명령 입력 - blacklist 명령 입력 - exit 명령 입력")
-    void customerCommandTypeBlacklist() {
+    @DisplayName("성공: customer 명령 입력")
+    void commandTypeCustomer() {
         //given
         given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.BLACKLIST, CustomerCommandType.EXIT);
 
         //when
         consoleMenu.runClient();
 
         //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(customerController).should().findBlacklistCustomers();
-    }
-
-    @Test
-    @DisplayName("성공: customer 명령 입력 - create 명령 입력 - exit 명령 입력")
-    void customerCommandTypeCreate() {
-        //given
-        given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.CREATE, CustomerCommandType.EXIT);
-
-        //when
-        consoleMenu.runClient();
-
-        //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(customerController).should().createCustomer();
-    }
-
-    @Test
-    @DisplayName("성공: customer 명령 입력 - list 명령 입력 - exit 명령 입력")
-    void customerCommandTypeList() {
-        //given
-        given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.LIST, CustomerCommandType.EXIT);
-
-        //when
-        consoleMenu.runClient();
-
-        //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(customerController).should().findCustomers();
-    }
-
-    @Test
-    @DisplayName("성공: customer 명령 입력 - update 명령 입력 - exit 명령 입력")
-    void customerCommandTypeUpdate() {
-        //given
-        given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.UPDATE, CustomerCommandType.EXIT);
-
-        //when
-        consoleMenu.runClient();
-
-        //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(customerController).should().updateCustomer();
-    }
-
-    @Test
-    @DisplayName("성공: customer 명령 입력 - delete 명령 입력 - exit 명령 입력")
-    void customerCommandTypeDelete() {
-        //given
-        given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.DELETE, CustomerCommandType.EXIT);
-
-        //when
-        consoleMenu.runClient();
-
-        //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(customerController).should().deleteCustomer();
-    }
-
-    @Test
-    @DisplayName("성공: customer 명령 입력 - help 명령 입력 - exit 명령 입력")
-    void customerCommandTypeHelp() {
-        //given
-        given(console.inputInitialCommand()).willReturn(ConsoleCommandType.CUSTOMER);
-        given(console.inputCustomerCommandType())
-                .willReturn(CustomerCommandType.HELP, CustomerCommandType.EXIT);
-
-        //when
-        consoleMenu.runClient();
-
-        //then
-        then(console).should().inputInitialCommand();
-        then(console).should(times(2)).inputCustomerCommandType();
-        then(console).should(times(2)).printCustomerCommandSet();
+        then(consoleCustomerMenu).should().runningCustomerService();
     }
 
     @Test
